feat(#89): F051-nestjs-file-convention-refactor 구현 업데이트
