+++ src/auth/auth-session.controller.ts
@@ -16,7 +16,7 @@ import type {
 } from './auth-session.service';
 import { Authenticated } from './decorators/authenticated.decorator';
 import { CurrentAuthSession } from './decorators/current-auth-session.decorator';
-import type { AuthenticatedRequest } from './auth-session-request.dto';
+import type { AuthenticatedRequest } from './dto/auth-session-request.dto';
 import {
   successResponse,
   type ApiSuccessResponse,
