+++ src/embedding/embedding.controller.svg.e2e.spec.ts
@@ -67,5 +67,5 @@ describe('EmbeddingController SVG (E2E)', () => {
     } finally {
       await app.close();
     }
-  });
+  }, 10000);
 });
